Update Program.cs
The whole idea is to prevent exhaustion of the InternalBufferSize property. This often occurs when events occur much faster than they are handled by the underlying functionality, since the buffer itself is only freed after the event handler returns; once the OnCreated call returns. n order to combat this, events will be moved to a separate queue to be processed from there, instead of on the FileSystemWatcher instance.